Add in backgrounding of the daemon and set up umask
We set up the umask to be 0 so that any process we spawn will have to
explicitly set their own, and we set up the daemon to be running within
the service directory. This way so long as the daemon is running nobody
can unmount the service directory out from under our nose.