Start of a wrapper around opendir and friends ...
This wrapper is meant to wrap some of the functionality of opendir and
friends. The reason why I want to wrap it is to provide iterators that
make it easier to deal with directories which are going to playing a
major part in the way services are going to be managed by dspawn.