Start building up the iterators
A lot of the difficult work has moved to the iterators, each time you
create a new iterator it will do the opendir and keep track of where it
is, rather than the base class. The base class has basically become an
empty shell that checks that you are trying to look at a directory and
then allows you to create iterators.

The reason this was done is so that if you have multiple iterators to
the same directory they don't keep overwriting each others results
because of the way that opendir/readdir/seekdir/and others work.

These itererators are kind of special though, there is no container
backing them, so that means that each time you start from the beginning,
you start from scratch with a completely new opendir/readdir cycle.
Hopefully the OS caching the results will be enough.