Add new errno exception class
This captures the errno and then uses it as a message when it gets
thrown this way we can start tracking when and where we are hitting
error numbers.

We may also want to start adding a way to get the error number
along-side of the message like what() does so that we can check later
down the exception handling chain what happened, although for now it is
strictly for errors that can not be "handled" and probably should cause
the app to crash or error out at least.

This class is now used everywhere there were TODO's in directory.cc.