Parse struct dirent*
We take the content of the struct dirent* and store it locally in
variables that exist for the lifetime of the object rather than what may
possibly be statically allocated memory.

Also add in two operators, one to check if the two directory_entries are
equal, the other to see if they are not equal (just calls the equality
operator for its operation and negates it). This allows it to be used to
see if two iterators are equal or not.

The only two fields we pull out are the name and the file number, as
those are the only two portable entries in a struct dirent*.