Move from make to scons
I had been attempting to stick with make in an attempt to keep
dependencies for building down to a minimum, however make was not nearly
as flexible as I had hoped and was causing me to write so many extra
entries that it was not worth my time and effort to continue to try
bending make to my will.

This also imports Google test into the tree and starts adding test
cases.